moved api and server actions to app root
