+++ apps/nextjs/src/app/(main)/organizations/[org_slug]/tournaments/[tournament_id]/(tournament_page)/@header/page.tsx
@@ -9,7 +9,7 @@ import type {
   OrganizationTournamentParams,
   OrganizationTournamentProps,
 } from "~/types";
-import { getSingleOrganizationSingleTournament } from "~/app/(main)/server-actions/organizations/tournaments/actions";
+import { getSingleOrganizationSingleTournament } from "~/app/server-actions/organizations/tournaments/actions";
 import OrganizationHeader from "~/components/organizations/organization-header";
 import { generateOrganizationTournamentsStaticParams } from "~/lib/organization-tournaments-static-params";
