+++ apps/nextjs/src/app/(main)/organizations/[org_slug]/tournaments/[tournament_id]/(tournament_page)/@registrations/page.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from "react";
 
 import type { OrganizationTournamentParams } from "~/types";
-import { getTournamentPlayers } from "~/app/(main)/server-actions/tournaments/actions";
+import { getTournamentPlayers } from "~/app/server-actions/tournaments/actions";
 import RegistrationsTable from "~/components/tournaments/registrations-table";
 
 export default async function TournamentRegistrationsPage(
