+++ apps/nextjs/src/app/(main)/tournaments/page.tsx
@@ -3,7 +3,7 @@ import { Suspense } from "react";
 
 import type { Organization, Tournament } from "@battle-stadium/db/schema";
 
-import { getOrganizationTournaments } from "~/app/(main)/server-actions/organizations/tournaments/actions";
+import { getOrganizationTournaments } from "~/app/server-actions/organizations/tournaments/actions";
 import { TournamentsTable } from "~/components/tournaments/tournaments-table";
 
 export const metadata: Metadata = {
