+++ apps/nextjs/src/app/api/cookies/user-id/route.ts
