+++ apps/nextjs/src/app/api/discord/interactions/route.ts
