+++ apps/nextjs/src/app/api/pokemon/team/validate/route.ts
