+++ apps/nextjs/src/components/organizations/partnered-organizations.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from "react";
 
-import { getPartneredOrganizations } from "~/app/(main)/server-actions/organizations/actions";
+import { getPartneredOrganizations } from "~/app/server-actions/organizations/actions";
 import OrganizationCard from "./organization-card";
 
 export default function PartneredOrganizationsSuspenseWrapper() {
